Move urljoin to a new utils.py package. Small fixes
