+++ backend/giscube_requests.py
@@ -3,27 +3,11 @@
 Package containing all the classes and utilities to communicate with the
 Giscube server.
 """
-import re
 
 from PyQt5.QtCore import QDir
 import requests
 
-
-def urljoin(base, *parts):
-    """
-    Joins the base and the parts separating them with a slash. Removes
-    duplicated slashes except if is part of the protocol (ie. https://).
-    :param base: Base url.
-    :type base: str
-    """
-    result = base
-    for part in parts:
-        result += '/'+part
-    result = re.sub(
-        r'(?<!/)(?<!:)/{2,}',
-        '/',
-        result)
-    return result
+from utils import urljoin
 
 
 class BadCredentials(ConnectionError):
